Kokkos,Tpetra: Change default graph/matrix offset type to LO
The default Kokkos::StaticCrsGraph<...>::size_type (type of row
offsets) is now the same as the type of column indices.  Ditto for
KokkosSparse::CrsMatrix and Tpetra::CrsGraph and Tpetra::CrsMatrix.
Had to do some awful metatype stuff to fix the FENL example, which
weirdly wants to create a Kokkos::StaticCrsGraph whose column
indices are unsigned[2] (???).