Tpetra::Distributor: Don't fill indicesFrom_ any more
@trilinos/tpetra

Tpetra::Distributor would always fill indicesFrom_ with 0, 1, ...,
totalReceiveLength_-1.  This was not only a waste of time, it would
unnecessarily make all reverse-mode communication take the slow path.
(Reversing a Distributor swaps indicesTo_ and indicesFrom_.  A
nonempty indicesTo_ tells Distributor to take the slow path.  If
indicesFrom_ is just 0, 1, ... totalReceiveLength_-1, then Distributor
does not need to take the slow path.)

Note that Epetra_MpiDistributor comments out the code that fills
indices_from_, and claims that doing so fixes reverse communication.

This commit changes Tpetra::Distributor so that it no longer fills
indicesFrom_ as above.