Tpetra::CrsGraph: Factor out maxDifference into new header
@trilinos/tpetra

Factor out Tpetra::Details::maxDifference (used in
CrsGraph::computeLocalConstants to compute the max number of local
entries) into a new header file.  Make it compute the min as well as
the max number of local entries (which we need to help set up the
sparse matrix-vector multiply kernel).  Rename it to
minMaxNumRowEntries.  Add a unit test for minMaxNumRowEntries.
Call the new function in CrsGraph::computeLocalConstants.